CR-3530 Fix for AME-3528: Mention CORS support in OpenAM release notes
In fact, as David Luna pointed out in the review, CORS support requires additional configuration in the deployment descriptor. So this patch also adds instructions to help with that, as reviewed. Those instructions are still likely to be useful, even if we include the configuration commented out in the default deployment descriptor file.

git-svn-id: https://svn.forgerock.org/openam/trunk@8987 0f4defcf-c51a-4c67-9f44-6fb5eba73c5d